Start splitting MozziGuts.cpp into a common core framework, and separate files for each main platform.
This is in the hope to increase maintainability and readability. Definitely seems to work out for STM32, for instance.

No testing done at all, so far.
